Researched and fixed ZC_DIALOG_NUMBERRANGE
The max length of the string is actually 128, it's followed by 2 ints,
that specifiy the min and max value. The min value will automatically be
inserted into the box. If the put in number is outside of that range you
can't send the result. If min and max are 0 you can send whatever number
you want, but the response is a byte, so it will overflow.