Disabled excessive packet logging
I think we're past the point where we need to inspect every single
packet. I've left the old logging code where it is, just in case, but
I've commented it, and I wrote a hex editor style ToString function for
Packet, that is used for unhandled packets. Looks nice and is easier to
copy, for get it into an editor.