Minor changes
Proposed changes for the last pull request. The coordinates are actually
easier to check without being a Position, and a TODO comment will make
sure we will stumble upon it again some time, as the IDEs pick it up.
The warning will ensure that we will notice and be able to debug
potential other values then the expected ones.