Minor change in packet handler try-catch
Changed the message to make it a little more clearer what happened
where, and added a full exception, incl. stacktrace. Afterwards the
server stops, assuming that we want all packet handlers to work, and
doesn't let the exception bubble up, because we don't need the vague
"failed to start" message after this.