Fixed hats still not appearing correctly everwhere
The packet celophi found was only for updating the state when it was
toggled, but it wasn't actually used anywhere else. I found three
suspicious bytes in AppearancePc though, where we had "pose" and
"teamId", which turned out to be the same hat visibility toggles. Hats
now work like you would expect, on yourself and on others. Only the
actual toggling isn't in yet.