Updated templates with properties
They all use the getProperties function now, which actually reads the
property bytes and lists keys and values. The one problem with this is
that not all properties are 4 bytes long (int/float), some are strings,
and some might be short or long, though we haven't seen any yet. This
means that 010 might fail to read a packet correctly if a property isn't
read correctly. In that case it needs to be added to the property type
switch in common.bt.