Added object property name getter template function
This didn't exist until now because it annoyed me how slow 010 is at
running the templates when this large switch case is included, but being
able to choose to see them, at the cost of some performance, seems
desirable. The property names will now be displayed in favor of the id
if the ObjectPropertyNames template is included.