Simplified compression in packets
Both the new and the old approach have advantages and disadvantages, and
room for errors from the user. However, the new one relies on only one
method call, instead of two that you can forgot about or call in an
incorrect order, it gives us a nice scope design for the "sub-packet"
with the delegate, and it allows for a simple passing of a "compress"
parameter, to change the compression on-the-fly with a single toggle,
which also means that you don't need a separate short to mark "no
compression" anymore. On top of all that it's also easier to maintain,
because we don't need special handling code in all Put methods anymore.