Removed job command
We have to rethink how this command is supposed to work, since you don't
just change your one job, you have a number a jobs, and your base job
always needs to exist in your jobs. We could treat it like a reset,
"remove all and change job to X", but that might be unexpected. We could
create add and remove commands, but that doesn't exactly handle the base
job. I'll hold off on a job command for now, until we've figured that
out, though we'll probably end up with 3 commands: reset, add, and
remove.