Refactored equip saving loop on char creation
Increasing a value one by one kinda works, but it feels wonky to me. A
small change and suddenly items go into the wrong slots. We're also
currently saving default items to the database, which we don't want
either, and adding a check for them would slightly complicate this loop.
Making it a property on the equip item makes more sense.