Updated monster class EXP
Based on the class EXP I'm receiving on iToS for Kepas and Leaf Bugs, it
seems like our class EXP were off by one decimal. Not sure how exactly
that's calculated, as Leaf Bug has a rate of 100 with a base of 20 for
class EXP, which should be 20 by my logic, not 2, but for now I've
divided all class EXP by 10, that works for the aforementioned monsters.