Added Character.ClassLevel
While this gives us the correct class level on the server, something is
wrong on the client. As soon as class level 15 is reached, the class
level display reverts back to level 1, with 0% EXP. I'm *pretty* sure
that doesn't happen on officials, and it's not the class EXP, I'm
sending the same values.