Added proper ReadArray and ReadList data methods
The purpose of the extension methods is to actually read and return values from JSON objects, in a way that is natural for a "reader" class in C#. Having a method just return a JToken cast to JArray doesn't improve the code a lot, since you still need to check existence and do the casts. These new methods handle that automatically.