Refactored item database
Previously, we used a category on the item data to determine the inventory index of an item. This worked pretty well because the categories used indices in increments of 5000 and we could simply use the category enum as a multiplier. Now, however, there are way more categories, sub-categories, and they don't come in simple 5k increments anymore either. I added the item type and group to the data, which are now used on launch to determine an item's category. Together with the InvBaseDb, this data can be used to determine the index range for every item. The category assigment code is not complete yet (see ItemDb), but it's a start and can be extended as we go along.