Improved combat entity generalization
Added ICombatEntity.CanAttack, to determine whether an entity is able to attack another on a case to case basis, and added ICombatEntity to Character, as they will be combatants. GetAttackableEntitiesInRange should maybe be moved to the Character/Monster classes, that would be cleaner, but since they'll be the same, maybe we should have a Combatant base class that they can use. It's stupid that they differentiate between characters and monsters >_>