Fixed minor issue in common.bt
Extra (bin) and size were backwards. Looking at this, we should probably read those bytes during the packet handling if they're a fix part of all client packets now. It seems unlikely that IMC added 12 bytes to every single packet, so it's probably part of the header, and we don't want to explicitly read it in every single handler.