Made TakeDamage parameters non-optional
With the optional attack index it wasn't clear whether the 0 that was passed to TakeDamage was the intended value for the damage visibility, or an explicit setting of the index before the visibility was added. Replaced with explicitly setting both parameters in full.  However, TakeDamage is getting a lot of information here, that it doesn't need to deal damage, that is just used for sending packets. Maybe it would be better to handle the packet sending outside, since the number of parameters will only increase, when we find new values in the hit packets, that we want to pass through from the skill handlers to the hit packets that are sent from TakeDamage.