Updated CZ_HEARTBEAT
ToS adds some padding to packets, to get to a multiple of 8, for their encryption. To be on the safe side, we only truncate these "garbage" bytes in fixed size packets in Pale, but not for dynamic packets. Instead, the number of presumed garbage bytes is displayed above the packet data. For this packet's template, the garbage bytes were counted as data, which was incorrect.