Added animation id data
Threre are >14000 animation ids that we need to be able to use, and while putting all of those into a static class or enum would've technically worked, the names were tricky to convert to valid symbol names. Having all of those also doesn't help much with Intellisense, since there are so many with similar names, that you will never find the right now. Instead, we're going to use a curated list of names in the core, as we are for other string-based values, and then look up the ids in the database.