Added experimental client script system
Scripts are sent to clients on login and on reload. For now, it simply sends all scripts found in either the system or user client scripts folders, treating them as an extension of the core, with the scripts making changes so the core can function as intended. Since the scripts make changes to the vanilla behavior of the client, it might be better to think of them as custom scripts, but for now we'll go with this and see where it leads us.