Added item monster data
The client generates the "monsters" that represent dropped items at run-time, with a sequential id. We could technically do this at run-time as well, but having a list makes debugging much easier. Let's use this for the moment and reevaluate at a later time. For reference: To get the monster ids, the client loads all item data, then sorts by class id, and then iterates through the items, createing monster with the id 800000+i, 800000 being the id of Default_Item in the monster data. If an id already exists, the next one is used. Afterwards the recipe items and monsters are generated, the monster id being Default_Recipe+i in this case.