Improved Position and Direction calculations
Also adds some unit tests. Directions can be a little confusing in ToS, since 0° is down when calculating a direction from it, but when it comes to actual angle values, they treat them normally. Additionally, the camera is at an angle on many maps for some reason, making it difficult to tell up from down. Maybe it would make more sense to treat the "ToS angle" as the special case.