Fixed drops getting picked back up immediately
If a monster drop was picked up and dropped again while the loot protection was still active it woud get picked up again immediately, because the loot protection check caught it. This moves that check to the bottom, so it can't be triggered accidentally, adds clearing the protections on pick up, and adds a delay before players can pick their items back up.