Added stat ratio bonuses
With this, the client-side preview calculations for the attack properties, when statting, should match the server calculations for now. We might want to consider using client-side scripts to modify the stat calculation functions, to adjust them to the server behavior. That way, we don't have to decode the client's behavior or worry about the calculations changing.