Added silver value to barrack character data
This makes the "Instanced Dungeon Usage" windows display the characters' silver, by adding a new field to the character data for the silver, which is saved together with the rest of the character data, to always get the latest silver amount on the login server. Technically we could get the silver amount by querying the item db, without duplicating the data, but this way is way simpler and arguably more efficient.