Fixed barrack team name change packets
We overcomplicated things a little and tried to read random memory data as actual data, which resulted in potential errors if this random memory data contained unexpected values. In reality, all that data appears to be just fixed-size strings. The "message" being 256 characters is just a guess, but it would make sense, and I really don't see what all those random bytes would be otherwise, in this fairly simple packet. This also fixes the team name change, after the initial setting, not using the TP dialog.