Improved character properties initialization
The setup for the auto-updating was moved to after the properties were loaded, because we can't rely on the properties coming in in any specific order, at least for now. We first need all properties, to get the proper initial and maximum values, then we can calculate the properties and set up the auto-updating.