Improved map entity management
Combat entities are now additionally added to a combined list of characters and monsters, for cases where we need to get entities from either pool. A ICombatEntity check was added in case we want to split up the classes more, which might be especially interesting for Monster, because items and props definitely aren't combat entities, but do fall under the Monster umbrella.