Added BuffCollection.Start
Simple method for starting buffs. Also modified the buff adding slightly. It doesn't feel quite right to me, because we're passing in Buff objects that don't necessarily get used. This is different from, say, skills, where the added skill becomes the skill and that's it. The new Start method feels a little more natural, because it just does what you want to do: start a buff.