Minor improvement in wander routine
The AI would sometimes not move because the random position was too close to its current position. The current wandering isn't well done yet, but it seems like there's a problem with monsters walking into props. Limiting wandering to an area around the initial position should minimize this for the moment.