Added server list
This adds a basic server class list we can use to manage the servers and use for proper look ups outside of db code, although it is still relying on the db data for now. The zone traffic packets now use this list and also properly list all available zone/channel servers. And to make that work I also added actual handling for the map ids. If they're set to "all" in the server db, it sets the zone server to handle all maps that exist in the map db. Otherwise, a list of explicit map ids is expected.