Created PropertyTable class
This is going to be the dedicated look up table for all things property ids. We could've used the db itself for this, but it's more convenient to have a globally accessible table, akin to an enum. We could also auto-generate the content, but loading it from a file makes it a little more flexible.