Replaced property system
Created a new generic-typed base system for variable management in Yggdrasil and used it to recreate the property system. The new one is similar on the surface, but its reusability is much better and we'll be able to use the base for the scripting variables as well. Additionally the properties now work with strings instead of integer ids. Not only weren't we able to rely on these ids never changing (which they already did multiple times in the past), we also wanted to be able to access shared properties in different namespaces with one call. For example, requesting the property "HP" should be possible without differentiating between "PC.HP" and "Monster.HP".