Moved "content" scripts to a new sub-folder
We'll want "content packs" to be easily toggleable, so we'll group them together in folders. They'll consist of NPCs, warps, shops, and whatever else could be considered content for the game. If someone doesn't want to use one set of content scripts, they can simply disable it and plug in their own pack. That's the idea so far, we'll see how it goes.