Moved warp scripts to content
Warps are arguably not integral to the game, but a matter of how a dev decides to connect the maps. If someone wanted to rearrange the world to fit their vision, they'd want to disable most, if not all, of the existing warps. As such, I believe they should be part of the content.