Updated monster stats
It seems like quite a few monsters use a new map-based stat override that changes their stats when they're spawned on specific maps. For example, Kepa has 118 HP on Melia and 660 on ToS. Other stats, like ATK and DEF are also affected by this. This commit changes all stats to the overrides for monsters that only have one entry in them, such as Kepa, which only spawns naturally in West Siauliai.