Changed Kepa properties for test spawns
Since Kepas only spawn in one location of the world, where their properties are fixed, we could use those fixed values as the default ones in the database. However, we would be making quite a few assumptions about the data at that point, because we don't really know which monsters might spawn under which circumstances in parts of the world we don't have data about. As such, I reverted the db stat changes and we're going to go the official route for the moment. I dislike that this is the way it's "meant to be" for the latest data, but I don't mind having this option, and we can decide for ourselves how we're going to define the stats for customized monsters and spawns.