Extended AI
Added functions to support build-up of hate, setting up actions to execute while routines are running, parameters to move routines to not wait until the entity reaches their destination, and a simple, experimental AI that makes use of all of it. Consider this a preview, as this system is a new approach for me when it comes to AIs and a few things are still in flux while I figure out how to utilize it to its full potential.