Added tendency support to Mob and AiScript
Regardless of what faction monsters belong to, they need an aggressive tendency to aggro players, and the game defines this tendency on a spawn-by-spawn basis. This system does feel janky, but it's what we need to be able to accurately emulate the game's aggro behavior without creating dedicated monsters or AIs for every eventuality. Personally, I would prefer the AIs to control the aggressiveness, and users can do that if they'd like, overwriting the tendency from the AI script.