Refactored skill handlers
Removed the handler interfaces and replaced them with dedicated ones for each known type we had so far. This is going to be a WIP as we figure out more about the skills. This commit also changes the handler parameters to use ICombatEntity instead of Character in preperation of mobs using them as well. Finally, this removes the default handlers for the time being, because it makes it more difficult to determine what's actually being handled and what goes through default handlers.