Removed default StopMove in BasicMonster.Idle
Added it as a quick way to stop when entering Idle from wherever, but I don't particuarly like catch-all calls like that. We'll see how I think about it after I stared at the new additional routine for a while, which we now need to accomplish stopping before going to Idle.