Increased default respawn delays
The game defines most spawners with a respawn delay of 0, but that means instantaneous respawn, which not only feels very wrong, it also doesn't match what we see in the game. To support this kind of data in a sensible way we're going to enforce a minimum respawn delay of 3 seconds, which is still very fast, and a max of triple that, to make it feel less mechanical.