Improved SkillHandlerAttribute
The attribute can now take a skill type and a name pattern to get the ids of all skills that match these arguments. Unfortunately attributes can't take predicates as arguments, so this will have to do. It will help us assigning handlers to series' of skills, such as all Mon_*_Attack1 skills.