Added self-harm check to CanAttack
It's a fun idea to allow friendly fire, including hitting yourself, but it's not going to be the default behavior, and there's also some caveats. For example, skills like Onion_Attack1 and Thrust use the caster's position as the origin, which may include them in the splash area by default. We'd have to make some decisions which skills to enable this for or how else it's going to work, because neither of these particular skills should really hurt the user.