Added HP priority to HitInfo
Required to update HP accurately via hit info packets. I do wonder if we really need the HP priority, as HP updates seemed to work fine in the past when we simply left it at 0, but this functionality was presumably added for a reason, and it might not hurt to do it properly.