Refactored monster properties
Moved the movement calculation to a script, added scriptable functions for all other properties as well, and added a new way to override monster properties by way of an Overrides property inside MonsterProperties. This is a dumb property value list that the calculation functions can check for overrides. Using calculated properties was a necessary change because monster properties can be affected from buffs and other sources as well, and there was no way around adding a method to define overrides because not all properties have built-in override properties.