Refactored skill hits
Thinking about how we want to calculate and apply damage modifiers such as crits, and how we'd probably not want to do that in every single skill handler, I decided to add another level to the skill hit handling. Prior to this, we used a general purpose damage calculation function, now there's a new method we call that not only returns the damage, but also additional information that the core can use to set up the (Skill)HitInfo. By sticking with the one-call design we can handle most, if not all, skills with one function call, and making this new function scriptable means we'll be prepared for potentially needing more of these that can easily be swapped, and we'll keep it as customizable as possible.