Added skill handler for Liberate
The buff required a small change to the AI hit event, since we need a reference to the attacker in the AI script. I don't know how happy I am with an explicit check for that one buff there, but I don't think there's a property for "threat". We could also introduce a getter on entities that can return the threat one entity feels towards another, but then the check is there, and we probably don't need this information anywhere else. I'll just leave it like this for now.