Fixed potential desync of cooldowns
Cooldowns are updated and removed on the heartbeat timer, so they're only updated every ~250ms. During that time it's technically possible to send a new request for something, such as using a skill, which would then fail because the server still has the action on cooldown. To solve this potential issue we need to check the actual times instead of whether the cooldown exists.