Added quest system client scripts
Modifies the client's quest code to ignore the game's quests and instead makes it list the ones we send from the server. Instead of going with a mostly custom system we could've tried to just insert custom quests into the existing pipeline, but the game's system is so inflexible and unwieldy that true customization would've been tricky. This goes somewhat against the spirit of emulation, but I do believe we'll have more fun with this custom system.