Updated AI Follow method
Added a matchSpeed parameter to choose whether the entity gets the same speed assigned as the target, added a warp to the target if it's too far away, and made it so the the AI would freeze once the target left for another map. This method is frankly getting a bit unwieldy and should be cleaned up, but for now the behavior should be matching the expected one closely.