Improved AddSpawner min amount
Looking around the game, the spawns appear way overpopulated, leading me to the conclusion that we probably do want flex spawning for client-based spawns as well. We'll set it to 50% for now, which honestly still seems a bit much. Though we have already seen this behavior of setting the min amount behind the scenes confuse users, so what we might want to do instead is actually define the min amount in our spawner calls, and make up numbers for the ones based on client data.