Simplified spawner shortcuts
Removed the property override overloads. These could potentially be useful in certain situations, but they aren't going to be used very often on a per-spawner basis. Instead I made the property on MonsterSpawner public, so one can set the overrides directly on the spawner. Also added a new overload that takes a map and all spawn delays.