Refactored friend accepting methods
Friends require a similar treatment as accounts. We need to be able to get them even if the associated users are offline and we need to update them on connected accounts as necessary. There's no real way around that, though the systems for it are subject to change. (Changes untested)