Updated communicator request handling
Now with more responsibilities and no additional features :+1: Requests can be sent as before, but they're not using an extension method anymore. The downside is that requests need to be specifically handled now and responses need to be wrapped in a response message. On the plus side, these requests will reliably return the response to a specific request now.