Added last login to social user
Friends need access to it without relying on the account, so the social users get their own login time. This really isn't a terrible idea though, because we might want to create a separate messenger-like app that connects to the social server, in which case we would log in with the user, but not the game account. The commit also includes setting the user reference on the friends.