Fixed rollback on warp (experimental)
As discussed on Discord, there is a potential issue with warping where a client might attempt to reconnect before we're done saving a character, in which case they would load old data and essentially experience a rollback, though the client would load the intended target map. What this looks like is that the character is dropped in a seemingly random position on the target map, but with the monsters and NPCs from the previous map still around. Essentially, the fix to this problem is to make sure we save the character before letting the client load outdated data. We could do this in different ways, but for simplicity's sake I've added a save in the warp finalization method in Character for now. It now saves the character and the account before instructing the client to load and connect to the target map (server). It also sets a flag to remember we just saved, so we don't save a second time in the connection closed handling.