Pets now attack, some logic reworked
A fairly significant rework of the blue orb logic.  The most significant change is that the master now stores a link to its blue orb summon and gets rid of it when warping / summoning a new one rather than having the pet try to do this, which is far cleaner and simplifies a lot of the code.

The pets also now have code to get aggro events when the master attacks or is attacked, which they use to trigger their attacking logic.

BlueOrb variables are also now named BlueOrb rather than just Orb in case RedOrbs ever need their own variables.