Improved TakeDamage methods
We were returning on death, which was somewhat unnecessary because that's what happened right afterwards anyway. Then we added the alert after the kill return, which meant we weren't getting alerts for kill hits. I think that we do want those though, since it allows the AI to react to deaths.