Fixed script projects
Since core projects no longer rely on a fixed list of project files and instead include everything in their folder, the barracks and zone script projects had to be moved a level down to their own sub-folders. This required a couple changes to the projects and the loading, but the only actually important information is that this is a breaking change in regards to script loading, as the user scripts folder needs to also be restructured. However, in most cases this should only require you to move your custom scripts one level down, into a "zone" folder. Examples: "user/scripts/zone/scripts.txt", "user/scripts/zone/my_npc.cs", referenced as "my_npc.cs".