Fixed loading of currently unavailable quests
Our quest loading code was failing when trying to load quests that are not currently or no longer available, due to the script not being loaded for example. The code was changed to not load quests that aren't available and to skip them during the saving process, which makes them retain their state in the db. When the quest script is available again they will be loaded as normal, so the player can continue where they left off.