Reverted some job exp changes
The EXP values of Rank 4 Levels 20 and above are a bit tricky to interpret because of floating point imprecision, but the values we had before appear to be correct. Or rather, they seem to match the client's interpretation. Tested by setting specific job EXP values and checking when the client would switch to the new level.