Update to templates
As mentioned in the previous PR for this feature, there were some documentation changes requested.
Note that MapRevealData does not include a second float, attempting to get another float gives the end of buffer error.  It seems the extra float sent while sending the data to the client is just a spacer.