Refactored AiScript.CanTarget
Changed it to be a method dedicated to checking whether an entity can be hated instead of "targeted", which was a bit vague and introduced new functionality that could've resulted in changed behavior. Also added a method for checking hate accumulation to replace the existing cloaking buff check. These two methods should hopefully provide a simple way to quickly disabled hate and hate gain under certain circumstances.