Improved Hold debuff
Added a counted hold to movement component that we can check for in the properties to fix the speed to 0. This is probably not how officials do it, but fixing the speed to 0 is an elegant solution to get this to apply to monsters and characters at once, and by counting the holds they can't be released accidentally. Though the official behavior should be researched some time.