Improved Full Draw's Link application
Moved the relevant code to a buff handler to keep the skill handler and combat calculations a little more concise, while also grouping the Link code together in one place. And while Full Draw appears to be the only skill to use this feature, this will also let us reuse it elsewhere easily.