Minor clean up of Oblique Shot
Among other things I removed comments that I felt only polluted the code. Our default assumption should be that the code is correct and as close to officials as it currently can be. Noting custom implementations is good, so we know about them, but if we clarify even for small things that they are correct according to X, the code will eventually be full of such comments.