Added TakeSimpleHit extension method
There will likely be a ton of cases where we need to apply a simple hit to a target, like from debuffs, and with this extension we don't have to repeat that pattern every single time. Time will tell if this was the right choice, but I'm pretty sure we want some common method for this.