Simplified HitInfo
If the HitInfo object won't always have a reference to the skill, we might as well remove that reference, because we won't be able to rely on it being there. We can then initialize the hit info with a skill id., and we can also remove the skill parameter from the sender, because that information is found in the hit info. I'll leave the attacker and target for now though, in case we need control over these. The actual benefit of this change is that the caller won't have to concern themselves with what they pass to HitInfo, whether they need a skill, or pass null, or whatever else.