Simplified collections code
Switched from using item class names internally to class ids, so we don't have to mess around with the item data as much. Also turned collection progress into a public class that can be passed around, so we don't have to awkwardly request the collection data from the component all the time, but can just iterate over the list of collections.