Minor style and clean up changes in Doppelsoeldner
Not many actual changes in this, but one noteworthy one is the addition of the skill modifier "FinalDamageMultiplier", which applies a multiplier after all other calculationes are done. While we can also do this inside the skill handlers, I feel it's more coherent to do all of this in the same place. It also saves us from doing it manually in every skill handler that needs it, and should we ever have to adjust it somehow, it only needs to be done in one place.