Added SetPropertyModifier for buffs
Does the same thing as UpdatePropertyModifier, but we might have a need to "Set" instead of "Add" modifiers in some cases, and even though we could've called Update then, it would've felt weird to call Update in the initial setting. Similarly, it feels more natural to call Update if you actually want to update. And finally, should we ever have a need for a distinction, both methods are there.