Updated character warp checks
Since we now support multiple zone servers, it doesn't make sense to check whether a given map exists on the current server. Rather, we need to either check if the data exists, or if there's a zone server with the map currently available. Switched to a data check for now.