Improved pad trigger subscription. Maybe.
It was a bit cumbersome that we had to check the types of the actors on every trigger subscriber, so I added a Subscribe method that does some automatic conversions behind the scenes to simplify combat-related pads slightly, which will presumably be most of them. See Shield Lob for examples of the new even handlers now supported, and note that this new type is now mandatory for formal pad handlers.