Added experimental combat buff "hook" functions
The idea is that buffs can inject their code into the combat handling and modify the results or modifiers by adding scriptable functions of their own, directly in the buff handlers, which are then called from these new functions if applicable. Is this design a good idea? I honestly don't have a single clue. There are certainly more performant options, but are they as flexible?