Updated SCR_CalculateDamage
We previously kept the damage value inside that function, treating it as the neutral function result, but not only do we kind of need the current damage in our events, like BeforeBonuses, it's also kind of pointless because we modify the skill hit result left and right anyway. Might as well use its damage property directly.