Improved buff and skill combat events
Realizing that we need two events per event, I added added just that. For example, BeforeCalc was split into AttackBeforeCalc and DefenseBeforeCalc, which may be called for either involved party depending on whether they're in the attacking role or defending. This is an important distiction, because we might otherwise apply a buff or skill twice if both parties have it and we call the same event for both. Twice the handlers twice the fun! Or something like that...