Reversed changes to SkillProjectile
The removal of the packet string check would leave us open to errors when a string doesn't exist, due to mistakes or typos. We do very much want those exceptions though. This change was made because there apparently are situations where 0 is used for a string id, but right now we're not actually using that anywhere. Should we actually need it, and we can't use an empty string as a packet string instead, we should add an option to pass in null, which might be a better representation of 0.