Added selection date to jobs
There appear to be potential issues when jobs are sent to the client "out of order", so we'll try to make sure the order is the one the jobs were selected in. Includes an SQL update that attempts to give viable selection dates to existing jobs, so the first and last jobs stay in order.