Improved MoveRandom AI routine
We used to simply check if the destination was valid, hoping for nothing to be in the way, but this sometimes caused monsters to walk through invalid terrain. Pathfinding was technically a solution to this, but the current path finding implementation doesn't produce great results for short paths without obstacles. So we'll go back to a straight line for now, but plot a line for a more accurate check of whether there is valid ground or not.