Improved IPF verification
Simplified the checksum setting, so the user only has to set the checksum to whatever a clean client sends. Also added an option to set the checksum conf option automatically, based on the checksum sent by devs logging in. This way you don't have to mess around with it manually, but just log in with an auth 99 account and a clean client.