Renamed skill handlers
I've been going back and forth on whether we should use the full skill class names on the files, because it just looks _so_ repetive, but I feel like this is still better for consistency. Not only is giving the file the same name as the class inside the most common convention, we're also doing this for buffs and abilities and it's just easier. Additionally, this allows reference searches for the file names based on the class name, which can be useful in some cases.