Improved buff duration setup
We were passing TimeSpan.Zero as the duration for some buffs to run them indefinitely, but that felt a bit cryptic and wasn't actually necessary in any of the cases we did it for. For improved readbility I added 2 "consts" to Buff though, to define default and infinite durations with more clarity.