Added CallSafe task helper
We're running a lot of our skill handler code on tasks for convenience, but this is also a bit risky, because an uncaught exception on such a task may crash the server. This commit adds a helper method that runs a method async as usual, but catches exceptions and logs them instead of letting them bubble up.