Added map-wide spawn area
We were lacking a way for users to quickly and easily define a map-wide spawn. While you could technically just create a giant rect that will undoubtedly span the whole map, specialized handling will be easier to understand and execute. A rect with a size of -1,-1, which can be created easily via `MapWide()`, will now receive special treatment and spawn a monster somewhere on the map.