Simplified MoveToAttack AI call
Moved the StopMove into the move routine, so the caller doesn't have to worry about that. Previously we had a manual MoveTo loop inside the AI, which made it necessary to stop the entity once we were in range, but with a specific move routine like this we don't really need that.